feat: Add numeric validations for path parameters
This commit adds numeric validations for the `item_id` path parameter in the `/items/{item_id}` endpoint. The `item_id` is annotated with `Annotated[int, Path(title="The ID of the item to get", gt=0, le=1000)]` to ensure that it is a positive integer between 0 and 1000. This enhancement improves the reliability and accuracy of the API by validating the input for the `item_id` path parameter.